1.moved CreateGenesisBlock from model to service 2.Changed CreateGenesisBlock to need full filepath as param. 3.Changed ConfigFromJson to need not filename but filepath
