+++ blockchain/domain/service/block.go
@@ -1,15 +1,14 @@
-package block
+package service
 
 import (
 	"encoding/json"
 	"time"
 
-	"github.com/it-chain/it-chain-Engine/blockchain/domain/service"
 	"github.com/it-chain/yggdrasill/block"
 )
 
-func CreateGenesisBlock() (*block.DefaultBlock, error) {
-	byteValue, err := service.ConfigFromJson("GenesisBlockConfig.json")
+func CreateGenesisBlock(genesisconfFilePath string) (*block.DefaultBlock, error) {
+	byteValue, err := ConfigFromJson(genesisconfFilePath)
 	if err != nil {
 		return nil, err
 	}
