+++ blockchain/domain/service/blockConfig_test.go
@@ -1,14 +1,18 @@
 package service
 
 import (
+	"go/build"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
 
 func TestConfigFromJson(t *testing.T) {
-	_, err1 := ConfigFromJson("GenesisBlockConfig.json")
+	genesisconfPath := build.Default.GOPATH + "/src/github.com/it-chain/it-chain-Engine/.it-chain/genesisconf/"
+	rightFilePath := genesisconfPath + "GenesisBlockConfig.json"
+	wrongFilePath := genesisconfPath + "WrongFileName.json"
+	_, err1 := ConfigFromJson(rightFilePath)
 	assert.NoError(t, err1)
-	_, err2 := ConfigFromJson("WrongFileName.json")
+	_, err2 := ConfigFromJson(wrongFilePath)
 	assert.Error(t, err2)
 }
